[asan] Support for %z to Printf()
At the moment, asan internal Printf() uses %l modifier for printing
values of size_t and related types. This works, because we control
both the implementation of Printf and all its uses, but can be a
little misleading.

This change adds support for %z to Printf(). All callers that print
sizes and pointers as integers are switched to %zu / %zx.


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@124655 a55fe9e4-9db4-d849-0498-ce57b7db7bff