Add support for building the ASan instrumentation unit tests with the just-built Clang binary, and linking them against the just-built ASan runtime.
This is *very* brittle. I expect it will require tweaking, and I've
pro-actively disabled it on non-Unix builds and on cross-builds.

It is also currently missing dependency edges on GoogleTest header files
and a few other corner cases, but those can be fixed. This is the major
milestone of a mini-bootstrap-like build of the unittest.


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@129826 a55fe9e4-9db4-d849-0498-ce57b7db7bff