[nolibc] Move all platforms to internal_getpid.
Before, we had an unused internal_getpid function for Linux, and a
platform-independent GetPid function.  To make the naming conventions
consistent for syscall-like functions, the GetPid syscall wrapper
in sanitizer_posix.cc is moved to sanitizer_mac.cc, and GetPid is
renamed to internal_getpid, bringing the Linux variant into use.


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@150296 a55fe9e4-9db4-d849-0498-ce57b7db7bff