Revert "[sanitizer] Intercept sched_getaffinity."
This reverts commit r187788.

The test case is unreliable (as the test may be run in a situation in
which it has no affinity with cpu0). This can be recommitted with a more
reliable test - possibly using CPU_COUNT != 0 instead (I wasn't entirely
sure that a process was guaranteed to have at least one affinity, though
it seems reasonable, or I'd have made the change myself).


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@155675 a55fe9e4-9db4-d849-0498-ce57b7db7bff