Only build embedded darwin variants if the compiler supports them
We need to filter out architectures that the compiler hasn't been built to
target (most likely the ARM ones) before attemptint to build a version of
libcompiler_rt.

This can result in a completely empty library (e.g. soft-float doesn't have any
x86 variants), in which case we shouldn't even try the build


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@164114 a55fe9e4-9db4-d849-0498-ce57b7db7bff