[asan] Install asan_device_setup to bin/ when targetting Android.
asan_device_setup is a utility that prepares a device to run code built with
ASan. Essentially, it installs ASan runtime library into the system. For this
reason, it has to be at a predictable relative path from the runtime library
itself. We also plan to distribute this utility, packaged with runtime library
and maybe llvm-symbolizer, to the users.


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@168874 a55fe9e4-9db4-d849-0498-ce57b7db7bff