[ASan] Fix coverage behavior when a PC belongs to an unknown address range.
This happens, e.g., when coverage data is collected for a module which is then
dlclose()'d. Currently this causes CovDump() to ignore all PCs that are greater
than the unrecognized PC. In other words, unloading a module causes ASan to
silently ignore any coverage data for modules loaded at higher addresses.
Instead we should just skip the unrecognized PCs.


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@174567 a55fe9e4-9db4-d849-0498-ce57b7db7bff