InstrProf: XFAIL tests from r208940 on Linux
According to the buildbots, the new features for shared objects don't
work on ELF since it requires an -fPIC when building the profile
library.  XFAIL these tests for now.

It's possible we'll have to build two versions of the profile library on
Linux (one with -fPIC and one without), but it's not clear to me exactly
how to resolve this.


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@175136 a55fe9e4-9db4-d849-0498-ce57b7db7bff