[asancov] Write coverage directly to a memory-mapped file.
This way does not require a __sanitizer_cov_dump() call. That's
important on Android, where apps can be killed at arbitrary time.

We write raw PCs to disk instead of module offsets; we also write
memory layout to a separate file. This increases dump size by the
factor of 2 on 64-bit systems.


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@175780 a55fe9e4-9db4-d849-0498-ce57b7db7bff