[TSan] Equalize the behavior of __tsan_get_allocated_size and user_alloc_usable_size.
The former used to crash with a null deref if it was given a not owned pointer,
while the latter returned 0. Now they both return 0. This is still not the best possible
behavior: it is better to print an error report with a stack trace, pointing
to the error in user code, as we do in ASan.


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@178148 a55fe9e4-9db4-d849-0498-ce57b7db7bff