[UBSan] Introduce UBSAN_OPTIONS environment variable.
If UBSan is run in a standalone mode (w/o any other sanitizer), it
still uses functions from sanitizer_common, some of which depend on
the value of runtime flags. Allow to override the default values of these
flags with UBSAN_OPTIONS variable. In particular, UBSAN_OPTIONS=symbolize=0
can be used to turn off online symbolization.


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@179732 a55fe9e4-9db4-d849-0498-ce57b7db7bff