builtins: re-enable building assembly sources
Re-apply SVN r213684 which was reverted in SVN r213724 since it broke the
build bots.  Add a tweak to enable inclusion of the assembly sources in
standalone build as well.

Original commit message:

This patch address the PR20360. The CMake assembler build system
ignores the .S assembly files in builtins library build. This patch
fixes the issue.


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@179954 a55fe9e4-9db4-d849-0498-ce57b7db7bff