[asan] De-flake one test.
This change replaces an in-test timeout with an unconditional blocking wait.
It speeds up normal execution significantly at the cost of hanging up indefinitely
in case of a failure. This is a very specific regression test and we don't
expect any failures in the future.

Another approach ould be increasing the timeout to ~8 seconds, which seems too
much for a lit test.


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@183401 a55fe9e4-9db4-d849-0498-ce57b7db7bff