log-path_test testcase - make the invalid path more invalid
The ASan test/asan/TestCases/log-path_test.cc testcase uses /INVALID as an invalid path and expects that the program will not be allowed to create or write to that file. This actually is a valid writable path on one of my setups. Let's make the path more invalid.

Reviewed at http://reviews.llvm.org/D6727


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@189740 a55fe9e4-9db4-d849-0498-ce57b7db7bff