Revert r224736: "[Sanitizer] Make CommonFlags immutable after initialization."
We've got some internal users that either aren't compatible with this or
have found a bug with it. Either way, this is an isolated cleanup and so
I'm reverting it to un-block folks while we investigate. Alexey and
I will be working on fixing everything up so this can be re-committed
soon. Sorry for the noise and any inconvenience.


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@190101 a55fe9e4-9db4-d849-0498-ce57b7db7bff