tsan: don't initialize trace header in release mode
We are going to use only a small part of the trace with the default
value of history_size. However, the constructor writes to the whole trace.
It writes mostly zeros, so freshly mmaped memory will do.
The only non-zero field if mutex type used for debugging.

Reduces per-goroutine overhead by 8K.

https://code.google.com/p/thread-sanitizer/issues/detail?id=89


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@193779 a55fe9e4-9db4-d849-0498-ce57b7db7bff