Silence some CMake 3.3 dev warnings in compiler-rt
Fix a trivial instance of CMP0054 that came up on llvmdev.

The other warnings were CMP0057, which is about using the same file as a
MAIN_DEPENDENCY multiple times.  The old behavior hasn't been a problem
yet, so I silenced the warning and filed PR23595 to document the issue
if someone cares.


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@201697 a55fe9e4-9db4-d849-0498-ce57b7db7bff