tsan: implement suppressions for top frame only
The new suppression type is called "race_top" and is matched only against top frame in report stacks.
This is required for situations when we want to suppress a race in a "thread pool" or "event loop" implementation.
If we simply use "race:ThreadPool::Execute" suppression, that can suppress everything in the program.

Reviewed in http://reviews.llvm.org/D10686


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@204587 a55fe9e4-9db4-d849-0498-ce57b7db7bff