[asan] Add OS X 10.11's new dyld interposition support
On OS X 10.11 (which is currently a public beta), the dynamic linker has been improved so that it doesn't require the use of DYLD_INSERT_LIBRARIES in order for interposition/wrappers to work. This patch adds support of this behavior into ASan – we no longer need to re-exec in case the env. variable is not set.

Reviewed at http://reviews.llvm.org/D10924


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@205118 a55fe9e4-9db4-d849-0498-ce57b7db7bff