[asan] Remove special case mapping on Android/AArch64.
ASan shadow on Android starts at address 0 for both historic and
performance reasons. This is possible because the platform mandates
-pie, which makes lower memory region always available.

This is not such a good idea on 64-bit platforms because of MAP_32BIT
incompatibility.

This patch changes Android/AArch64 mapping to be the same as that of
Linux/AAarch64.


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@206976 a55fe9e4-9db4-d849-0498-ce57b7db7bff