[asan] Zero initialize sem_t in sem_init.
Old version of sem_init (GLIBC_2.0) fails to initialize parts of
sem_t that are used in sem_timedwait. This is fixed in GLIBC_2.1,
but since ASan interceptors downgrade sem_* to the oldest available
version, this can introduce bugs that are only present in sanitized
build. Workaround by zero-initializing sem_t in sem_init.


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@213121 a55fe9e4-9db4-d849-0498-ce57b7db7bff