[tsan] Handle libdispatch worker threads on OS X
On OS X, GCD worker threads are created without a call to pthread_create. We need to properly register these threads with ThreadCreate and ThreadStart. This patch uses a libpthread API (`pthread_introspection_hook_install`) to get notifications about new threads and about threads that are about to be destroyed.

Differential Revision: http://reviews.llvm.org/D14328


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@214990 a55fe9e4-9db4-d849-0498-ce57b7db7bff