Revert "[CMake] [macho_embedded] Check to make sure the compiler supports the architectures before generating build targets"
This commit reverts r252525. I was not really thinking about this fix properly. This doesn't work because it relys on try_compile checks which do a full compile & link. I'm going to put in a temporary solution as an interm step until we have a way to perform compiler checks without linking.


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@215460 a55fe9e4-9db4-d849-0498-ce57b7db7bff