[tsan] Fix stack_sync_reuse.cc test on OS X
The test relies on two variables in different frames to end up being on the same address. For some reason, this isn't true on OS X. This patch adds `__attribute__((aligned(64)))` to the variables, which actually makes the variables occupy the same address. This is still not a guarantee, but it's more likely to work (the test looks very fragile already).

Differential Revision: http://reviews.llvm.org/D14925


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@216753 a55fe9e4-9db4-d849-0498-ce57b7db7bff