When building without DIA SDK, don't set suppressions.cpp fail
This test requires llvm-symbolizer to be able to convert a stack
address into a function name.  It is only able to do this if the
DIA SDK was found at cmake time.  Add a lit feature for this,
and let the test depend on it.

See also discussion in D15363.


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@220967 a55fe9e4-9db4-d849-0498-ce57b7db7bff