[libprofile] Adjust dummy condition to be more forgiving.
On PS4, we have to fake environment variables by passing extra command
line arguments, so the dummy test `argc > 1` was failing.
The condition is just a dummy condition that the compiler can't fold
away, so the number is arbitrary as long as the condition is false.

Increase the number it compares against.


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@226288 a55fe9e4-9db4-d849-0498-ce57b7db7bff