[tsan] Remove long sleeps from fork tests
On one of our testing machines, we're running the tests under heavy load, and especially in the fork-based TSan tests, we're seeing timeouts when a test uses sleep(10), assuming that calling fork() on another thread will finish sooner than that. This patch removes a timeout and makes another one longer.

Differential Revision: http://reviews.llvm.org/D18476


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@227367 a55fe9e4-9db4-d849-0498-ce57b7db7bff