[tsan] Synchronize leaving a GCD group with notifications
In the patch that introduced support for GCD barrier blocks, I removed releasing a group when leaving it (in dispatch_group_leave). However, this is necessary to synchronize leaving a group and a notification callback (dispatch_group_notify). Adding this back, simplifying dispatch_group_notify_f and adding a test case.

Differential Revision: http://reviews.llvm.org/D21927


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@235706 a55fe9e4-9db4-d849-0498-ce57b7db7bff