test: Use %clangxx in objc++ test files
These test in this change are objc++, but are built using %clang, not %clangxx.
The reason this works is the driver has been adding -lc++ for sanitizer enabled
builds. By making these tests use %clangxx, they no longer depend on the driver
linking to c++.  Doing so will allow us to prevent overlinking of libc++ for
applications.


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@236128 a55fe9e4-9db4-d849-0498-ce57b7db7bff