[asan] Fix Win64 test portability issues
The OOM test should really only run on 32-bits, since it's hard to OOM
on x64.

The operator_array_new_with_dtor_left_oob tests need to account for the
larger array cookie on x64 (8 bytes instead of 4).

Use -std=c++14 in use-after-scope-capture.cc to avoid errors in the MSVC
2015 STL on Windows. The default there is C++14 anyway.


git-svn-id: https://github.com/joker-eph/llvm-project/trunk/compiler-rt@237381 a55fe9e4-9db4-d849-0498-ce57b7db7bff